swiftformat tests
