+++ Makefile
@@ -6,7 +6,12 @@ PLATFORM_WATCHOS = watchOS Simulator,name=Apple Watch Series 7 (45mm)
 
 ROOT_DIR := $(dir $(realpath $(lastword $(MAKEFILE_LIST))))
 
-.PHONY: submodules test
+.PHONY: submodules test dev clean
+
+dev:
+	./scripts/bootstrap
+	make clean
+	make submodules
 
 clean:
 	rm -rf "$(ROOT_DIR)Sources/secp256k1/libsecp256k1"
