+++ Tests/K1Tests/TestCases/ECDSA/ECDSASignatureTests.swift
@@ -1,47 +1,40 @@
-//
-//  File.swift
-//  
-//
-//  Created by Alexander Cyon on 2022-02-03.
-//
-
+import CryptoKit
 import Foundation
 import K1
 import XCTest
-import CryptoKit
 
 final class ECDSASignatureTests: XCTestCase {
-    func testECDSADeterministic() throws {
-        let alice = K1.ECDSA.NonRecoverable.PrivateKey()
-        let message = "Send Bob 3 BTC".data(using: .utf8)!
-       
-        let signature = try alice.signature(forUnhashed: message)
-        
-        let isSignatureValid = alice.publicKey.isValidSignature(signature, unhashed: message)
-        XCTAssertTrue(isSignatureValid, "Signature should be valid.")
-    }
+	func testECDSADeterministic() throws {
+		let alice = K1.ECDSA.NonRecoverable.PrivateKey()
+		let message = "Send Bob 3 BTC".data(using: .utf8)!
 
-    func testECDSARandom() throws {
-        continueAfterFailure = false
-        let alice = K1.ECDSA.NonRecoverable.PrivateKey()
-        let message = "Send Bob 3 BTC".data(using: .utf8)!
-       
-        let requestedNumberOfSignatures = 1000
-        var signatures = Set<K1.ECDSA.NonRecoverable.Signature>()
-       
-        for i in 0..<requestedNumberOfSignatures {
-            let signature = try alice.signature(
-                forUnhashed: message,
-                options: .init(nonceFunction: .random)
-            )
-            
-            let isSignatureValid = alice.publicKey.isValidSignature(
-                signature,
-                unhashed: message
-            )
-            XCTAssertTrue(isSignatureValid, "Signature should be valid.")
-            XCTAssertEqual(signatures.count, i)
-            signatures.insert(signature)
-        }
-    }
+		let signature = try alice.signature(forUnhashed: message)
+
+		let isSignatureValid = alice.publicKey.isValidSignature(signature, unhashed: message)
+		XCTAssertTrue(isSignatureValid, "Signature should be valid.")
+	}
+
+	func testECDSARandom() throws {
+		continueAfterFailure = false
+		let alice = K1.ECDSA.NonRecoverable.PrivateKey()
+		let message = "Send Bob 3 BTC".data(using: .utf8)!
+
+		let requestedNumberOfSignatures = 1000
+		var signatures = Set<K1.ECDSA.NonRecoverable.Signature>()
+
+		for i in 0 ..< requestedNumberOfSignatures {
+			let signature = try alice.signature(
+				forUnhashed: message,
+				options: .init(nonceFunction: .random)
+			)
+
+			let isSignatureValid = alice.publicKey.isValidSignature(
+				signature,
+				unhashed: message
+			)
+			XCTAssertTrue(isSignatureValid, "Signature should be valid.")
+			XCTAssertEqual(signatures.count, i)
+			signatures.insert(signature)
+		}
+	}
 }
