+++ Tests/K1Tests/TestCases/ECDSA/ECDSAWycheproofASNDEREncodedSignaturesTests.swift
@@ -1,5 +1,4 @@
-// FROM: https://github.com/apple/swift-crypto/blob/main/Tests/CryptoTests/Signatures/ECDSA/ECDSASignatureTests.swift
-// commit: 53da7b3706ae6a2bd621becbb201f3d8e24039d6
+// swiftformat:disable strip
 
 //===----------------------------------------------------------------------===//
 //
@@ -14,54 +13,60 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 //===----------------------------------------------------------------------===//
-import Foundation
-import XCTest
-import CryptoKit
-@testable import K1
 
+// FROM: https://github.com/apple/swift-crypto/blob/main/Tests/CryptoTests/Signatures/ECDSA/ECDSASignatureTests.swift
+// commit: 53da7b3706ae6a2bd621becbb201f3d8e24039d6
+
+// swiftformat:enable strip
+
+import CryptoKit
+import Foundation
+@testable import K1
+import XCTest
+
+// MARK: - ECDSA_Wycheproof_ASN_DER_EncodedSignaturesTests
 final class ECDSA_Wycheproof_ASN_DER_EncodedSignaturesTests: XCTestCase {
-    
-    func testWycheProofSecp256k1_DER() throws {
-        let result: TestResult = try testSuite(
-            /* https://github.com/google/wycheproof/blob/master/testvectors/ecdsa_secp256k1_sha256_test.json */
-            jsonName: "wycheproof_ecdsa_verify_der",
-            testFunction: { (group: ECDSAWycheTestGroup<SignatureWycheproofDERTestVector>) in
-                
-                try doTestGroup(
-                    group: group,
-                    signatureValidationMode: .init(malleabilityStrictness: .accepted),
-                    hashFunction: SHA256.self,
-                    skipIfContainsFlags: .init(["MissingZero", "BER"])
-                )
-                
-            })
-        print("☑️ Test result: \(String(describing: result))")
-    }
+	func testWycheProofSecp256k1_DER() throws {
+		let result: TestResult = try testSuite(
+			/* https://github.com/google/wycheproof/blob/master/testvectors/ecdsa_secp256k1_sha256_test.json */
+			jsonName: "wycheproof_ecdsa_verify_der",
+			testFunction: { (group: ECDSAWycheTestGroup<SignatureWycheproofDERTestVector>) in
+
+				try doTestGroup(
+					group: group,
+					signatureValidationMode: .init(malleabilityStrictness: .accepted),
+					hashFunction: SHA256.self,
+					skipIfContainsFlags: .init(["MissingZero", "BER"])
+				)
+			}
+		)
+		print("☑️ Test result: \(String(describing: result))")
+	}
 }
 
+// MARK: - SignatureWycheproofDERTestVector
 private struct SignatureWycheproofDERTestVector: WycheproofTestVector {
-    
-    typealias MessageDigest = SHA256.Digest
-    typealias Signature = K1.ECDSA.NonRecoverable.Signature
-    
-    let comment: String
-    let msg: String
-    let sig: String
-    let result: String
-    let flags: [String]
-    let tcId: Int
-    
-    func messageDigest() throws -> MessageDigest {
-        let msg = try Data(hex: msg)
-        return SHA256.hash(data: msg)
-    }
-    func expectedSignature() throws -> Signature {
-        let derData = try Data(hex: sig)
-        let signature = try K1.ECDSA.NonRecoverable.Signature(derRepresentation: derData)
-        if self.result == "valid" {
-            try XCTAssertEqual(sig, signature.derRepresentation().hex)
-        }
-        return signature
-    }
-    
+	typealias MessageDigest = SHA256.Digest
+	typealias Signature = K1.ECDSA.NonRecoverable.Signature
+
+	let comment: String
+	let msg: String
+	let sig: String
+	let result: String
+	let flags: [String]
+	let tcId: Int
+
+	func messageDigest() throws -> MessageDigest {
+		let msg = try Data(hex: msg)
+		return SHA256.hash(data: msg)
+	}
+
+	func expectedSignature() throws -> Signature {
+		let derData = try Data(hex: sig)
+		let signature = try K1.ECDSA.NonRecoverable.Signature(derRepresentation: derData)
+		if self.result == "valid" {
+			try XCTAssertEqual(sig, signature.derRepresentation().hex)
+		}
+		return signature
+	}
 }
