+++ Tests/K1Tests/TestCases/Schnorr/SchnorrSignatureTests.swift
@@ -1,23 +1,14 @@
-//
-//  File.swift
-//  
-//
-//  Created by Alexander Cyon on 2022-02-03.
-//
-
+import CryptoKit
 import Foundation
 import K1
 import XCTest
-import CryptoKit
 
 final class SchnorrSignatureTests: XCTestCase {
-    
-    func testSchnorr() throws {
-        let alice = K1.Schnorr.PrivateKey()
-        let message = "Send Bob 3 BTC".data(using: .utf8)!
-        let signature = try alice.signature(forUnhashed: message)
-        let isSignatureValid = alice.publicKey.isValidSignature(signature, unhashed: message)
-        XCTAssertTrue(isSignatureValid, "Signature should be valid.")
-    }
-    
+	func testSchnorr() throws {
+		let alice = K1.Schnorr.PrivateKey()
+		let message = "Send Bob 3 BTC".data(using: .utf8)!
+		let signature = try alice.signature(forUnhashed: message)
+		let isSignatureValid = alice.publicKey.isValidSignature(signature, unhashed: message)
+		XCTAssertTrue(isSignatureValid, "Signature should be valid.")
+	}
 }
