+++ Tests/K1Tests/Util/ECSignature.swift
@@ -1,13 +1,11 @@
-//
-//  File.swift
-//  
-//
-//  Created by Alexander Cyon on 2023-03-19.
-//
-
 import Foundation
 import K1
 
+// MARK: - ECSignature
 public protocol ECSignature {}
+
+// MARK: - K1.ECDSA.NonRecoverable.Signature + ECSignature
 extension K1.ECDSA.NonRecoverable.Signature: ECSignature {}
+
+// MARK: - K1.ECDSA.Recoverable.Signature + ECSignature
 extension K1.ECDSA.Recoverable.Signature: ECSignature {}
