+++ scripts/pre-commit
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+DIR="$(dirname "$(readlink -f "$BASH_SOURCE")")"
+
+"$DIR/git-format-staged" --formatter "swiftformat stdin --stdinpath '{}'" "*.swift"
